+++ src/pages/RepoPage/CoverageOnboarding/GitHubActions/GitHubActions.spec.tsx
@@ -5,20 +5,22 @@ import { setupServer } from 'msw/node'
 import { Suspense } from 'react'
 import { MemoryRouter, Route } from 'react-router-dom'
 
+import { useFlags } from 'shared/featureFlags'
+
 import GitHubActions from './GitHubActions'
 
-const mockCurrentUser = {
-  me: {
-    trackingMetadata: {
-      ownerid: 'user-owner-id',
-    },
-  },
-}
+jest.mock('./GitHubActionsRepoToken', () => () => 'GitHubActionsRepoToken')
+jest.mock('./GitHubActionsOrgToken', () => () => 'GitHubActionsOrgToken')
+jest.mock('shared/featureFlags')
 
-const mockGetRepo = {
+const mockedNewRepoFlag = useFlags as jest.Mock<{ newRepoFlag: boolean }>
+
+const mockGetRepo = (hasOrgUploadToken: boolean | null) => ({
   owner: {
     isCurrentUserPartOfOrg: true,
-    orgUploadToken: '9e6a6189-20f1-482d-ab62-ecfaa2629290',
+    orgUploadToken: hasOrgUploadToken
+      ? '9e6a6189-20f1-482d-ab62-ecfaa2629290'
+      : null,
     repository: {
       private: false,
       uploadToken: '9e6a6189-20f1-482d-ab62-ecfaa2629295',
@@ -28,7 +30,7 @@ const mockGetRepo = {
       oldestCommitAt: '',
     },
   },
-}
+})
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -66,126 +68,43 @@ afterEach(() => {
 afterAll(() => server.close())
 
 describe('GitHubActions', () => {
-  function setup() {
+  function setup(hasOrgUploadToken: boolean | null) {
+    mockedNewRepoFlag.mockReturnValue({ newRepoFlag: true })
+
     server.use(
       graphql.query('GetRepo', (req, res, ctx) =>
-        res(ctx.status(200), ctx.data(mockGetRepo))
-      ),
-      graphql.query('CurrentUser', (req, res, ctx) =>
-        res(ctx.status(200), ctx.data(mockCurrentUser))
+        res(ctx.status(200), ctx.data(mockGetRepo(hasOrgUploadToken)))
       )
     )
   }
 
-  describe('step one', () => {
-    beforeEach(() => setup())
-
-    it('renders header', async () => {
-      render(<GitHubActions />, { wrapper })
-
-      const header = await screen.findByRole('heading', { name: /Step 1/ })
-      expect(header).toBeInTheDocument()
-
-      const repositorySecretLink = await screen.findByRole('link', {
-        name: /repository secret/,
-      })
-      expect(repositorySecretLink).toBeInTheDocument()
-      expect(repositorySecretLink).toHaveAttribute(
-        'href',
-        'https://github.com/codecov/cool-repo/settings/secrets/actions'
-      )
+  describe('when org upload token is available', () => {
+    beforeEach(() => {
+      setup(true)
     })
 
-    it('renders body', async () => {
+    it('renders GitHubActionsOrgToken', async () => {
       render(<GitHubActions />, { wrapper })
 
-      const body = await screen.findByText(
-        /Admin required to access repo settings > secrets and variable > actions/
+      const githubActionsOrgToken = await screen.findByText(
+        'GitHubActionsOrgToken'
       )
-      expect(body).toBeInTheDocument()
-    })
-
-    it('renders token box', async () => {
-      render(<GitHubActions />, { wrapper })
-
-      const codecovToken = await screen.findByText(/CODECOV_TOKEN=/)
-      expect(codecovToken).toBeInTheDocument()
-
-      const tokenValue = await screen.findByText(
-        /9e6a6189-20f1-482d-ab62-ecfaa2629295/
-      )
-      expect(tokenValue).toBeInTheDocument()
+      expect(githubActionsOrgToken).toBeInTheDocument()
     })
   })
 
-  describe('step two', () => {
-    beforeEach(() => setup())
+  describe('when org upload token is not available', () => {
+    beforeEach(() => {
+      setup(false)
+    })
 
-    it('renders header', async () => {
+    it('renders GitHubActionsRepoToken', async () => {
       render(<GitHubActions />, { wrapper })
 
-      const header = await screen.findByRole('heading', { name: /Step 2/ })
-      expect(header).toBeInTheDocument()
-
-      const gitHubActionsWorkflowLink = await screen.findByRole('link', {
-        name: /GitHub Actions workflow/,
-      })
-      expect(gitHubActionsWorkflowLink).toBeInTheDocument()
-      expect(gitHubActionsWorkflowLink).toHaveAttribute(
-        'href',
-        'https://github.com/codecov/cool-repo/tree/main/.github/workflows'
+      const githubActionsRepoToken = await screen.findByText(
+        'GitHubActionsRepoToken'
       )
-    })
-
-    it('renders yaml section', async () => {
-      render(<GitHubActions />, { wrapper })
-
-      const yamlBox = await screen.findByText(
-        /Upload coverage reports to Codecov/
-      )
-      expect(yamlBox).toBeInTheDocument()
-    })
-  })
-
-  describe('step three', () => {
-    beforeEach(() => setup())
-    it('renders first body', async () => {
-      render(<GitHubActions />, { wrapper })
-
-      const body = await screen.findByText(/After you committed your changes/)
-      expect(body).toBeInTheDocument()
-    })
-
-    it('renders second body', async () => {
-      render(<GitHubActions />, { wrapper })
-
-      const body = await screen.findByText(/Once merged to the/)
-      expect(body).toBeInTheDocument()
-    })
-
-    it('renders status check image', async () => {
-      render(<GitHubActions />, { wrapper })
-
-      const img = await screen.findByRole('img', {
-        name: 'codecov patch and project',
-      })
-      expect(img).toBeInTheDocument()
-    })
-  })
-
-  describe('ending', () => {
-    beforeEach(() => setup())
-    it('renders body', async () => {
-      render(<GitHubActions />, { wrapper })
-
-      const body = await screen.findByText(/How was your setup experience/)
-      expect(body).toBeInTheDocument()
-
-      const bodyLink = await screen.findByRole('link', { name: /this issue/ })
-      expect(bodyLink).toHaveAttribute(
-        'href',
-        'https://github.com/codecov/Codecov-user-feedback/issues/18'
-      )
+      expect(githubActionsRepoToken).toBeInTheDocument()
     })
   })
 })
